Switch to library-go resourceapply
As per openshift/machine-api-operator#541 and MAO commit 4e3ac1e, the
MAO switched from the deep-compare based CVO resourceapply library to
the library-go generation based library. So we follow suit here.

The generations structure is currently just stored in memory, which
means we will lose track of this on restart and re-apply the
deployment. This isn't a big problem, but eventually we should persist
the generations to a CR.